+++ application/models/stream_model.php
@@ -10,7 +10,9 @@ class Stream_model extends CI_Model
   public function __construct()
   {
     parent::__construct();
-    $this->load->driver('cache', array('adapter' => 'file'));
+    if($this->load->is_supported('apc')) { $this->load->driver('cache', array('adapter' => 'apc')); }
+    elseif($this->load->is_supported('memcached')) { $this->load->driver('cache', array('adapter' => 'memcached')); }
+    elseif($this->load->is_supported('file')) { $this->load->driver('cache', array('adapter' => 'file')); }
   }
 
   // -- here --